+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## 2.0.0
+ * STORM-1262: port backtype.storm.command.dev-zookeeper to java.
+ * STORM-1243: port backtype.storm.command.healthcheck to java.
+ * STORM-1246: port backtype.storm.local-state to java.
+ * STORM-1516: Fixed issue in writing pids with distributed cluster mode.
+ * STORM-1253: port backtype.storm.timer to java
+ * STORM-1258: port thrift.clj to Thrift.java
+ * STORM-1336: Evalute/Port JStorm cgroup support and implement cgroup support for RAS
  * STORM-1511: min/max operators implementation in Trident streams API.
  * STROM-1263: port backtype.storm.command.kill-topology to java
  * STORM-1260: port backtype.storm.command.activate to java